Fix Alert

+++ model/admin/product.php
@@ -129,21 +129,21 @@ if ($_SESSION['user_id'] != 'admin') {
               echo "<td><img src='../../view/produk/" . $row["foto"] . "' alt='foto' width='100'></td>";
               echo "<td>" . $row["harga"] . "</td>";
               echo "<td>" . ($row["diskon"] ? "Ya" : "Tidak") . "</td>";
-              echo "<td><a href='form/update_wisata.php?id=" . $row["id"] . "' class='btn btn-success'>Update</a></td>";
-              echo "<td><a href='../../controller/delete_wisata.php?id=" . $row["id"] . "' class='btn btn-danger'>Delete</a></td>";
+              echo "<td><a href='form/update_wisata.php?id=" . $row["id"] . "' class='btn btn-success' onclick='return confirm(\"Apakah Anda yakin ingin meng-Update Wisata?\");'>Update</a></td>";
+              echo "<td><a href='../../controller/delete_wisata.php?id=" . $row["id"] . "' class='btn btn-danger' onclick='return confirm(\"Apakah Anda yakin ingin meng-Hapus Wisata?\");'>Delete</a></td>";
 
               echo "</tr>";
             }
           } else {
             echo "<tr>
-            <td colspan='1'>Tidak ada data.</td>
-            <td colspan='1'>Tidak ada data.</td>
-            <td colspan='1'>Tidak ada data.</td>
-            <td colspan='1'>Tidak ada data.</td>
-            <td colspan='1'>Tidak ada data.</td>
-            <td colspan='1'>Tidak ada data.</td>
-            <td colspan='1'>Tidak ada data.</td>
-            <td colspan='1'>Tidak ada data.</td>
+            <td colspan='6' class='text-center'>Tidak ada data.</td>
+            <td colspan='0' class='d-none'></td>
+            <td colspan='0' class='d-none'></td>
+            <td colspan='0' class='d-none'></td>
+            <td colspan='0' class='d-none'></td>  
+            <td colspan='0' class='d-none'></td>
+            <td colspan='0' class='d-none'></td>
+            <td colspan='0' class='d-none'></td>
             </tr>";
           }
